副作用のあるimportを避けるように修正
サーバまで起動してしまったがそれは不本意、
サーバ起動用のコードだけserver-ts/index.tsに残して
型定義をserver-ts/app.tsに移動
