+++ server-ts/package.json
@@ -8,7 +8,7 @@
   },
   "exports": {
     ".": {
-      "types": "./src/index.ts"
+      "import": "./src/app.ts"
     }
   },
   "dependencies": {
